requirements.txt
Flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-bcrypt==0.7.1
flask-apscheduler==1.12.2
Werkzeug==2.0.1
SQLAlchemy==1.4.23
APScheduler==3.8.1
python-dotenv==0.19.0
gunicorn==20.1.0
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
markupsafe==2.0.1
pytz==2021.1
tzlocal==3.0.0
six==1.16.0
# If you're using MySQL, uncomment these:
# mysqlclient==2.1.1
# PyMySQL==1.0.2
